// src/models/ISkill.ts
export interface ISkill {
  _id: string;
  name: string;
  category: string;
  status: string;
  addSkill: string;
  tech: string;
  languageProgram: string;
  search: string;
  ategory: string;
  action: string;
  edit: string;
  delete: string;
  save: string;
  cancle: string;
  confirm1: string;
  confirm2: string;
  statusAdd: string;
  statusEdit: string;
  statusDelete: string;
  statusFail: string;
  active: string;
  inactive: string;
}